+++ activerecord/test/cases/explain_subscriber_test.rb
@@ -38,6 +38,13 @@ if ActiveRecord::Base.connection.supports_explain?
       end
     end
 
+    def test_collects_nothing_if_unexplained_sqls
+      with_queries([]) do |queries|
+        SUBSCRIBER.call(:name => 'SQL', :sql => 'SHOW max_identifier_length')
+        assert queries.empty?
+      end
+    end
+
     def with_queries(queries)
       Thread.current[:available_queries_for_explain] = queries
       yield queries